[Refactor] Generalize GRPO stepper into MixedPrecisionOptimizationStepper
Nothing in the stepper was GRPO-specific: autocast, GradScaler, gradient
accumulation and gradient clipping are generic training concerns. Move it
to torchrl.trainers.trainers next to DefaultOptimizationStepper, rename
it MixedPrecisionOptimizationStepper, and export it from torchrl.trainers
so any trainer can use it. GRPOTrainer keeps building it by default.

While moving:
- Drop the hardcoded "cuda" device: autocast and GradScaler now use a
  `device_type` argument that defaults to the device type of the
  optimizer's first parameter.
- Remove the unconditional gc.collect() after every optimizer step - an
  expensive global side effect that belongs to the calling script, not a
  library stepper.
- Make the loss-key error message loss-module agnostic.
