+++ torchrl/trainers/algorithms/__init__.py
@@ -9,7 +9,7 @@ from .a2c import A2CTrainer
 from .cql import CQLTrainer
 from .ddpg import DDPGTrainer
 from .dqn import DQNTrainer
-from .grpo import GRPOOptimizationStepper, GRPOTrainer
+from .grpo import GRPOTrainer
 from .iql import IQLTrainer
 from .offline_to_online import OfflineToOnlineTrainer
 from .on_policy import OnPolicyTrainer
@@ -23,7 +23,6 @@ __all__ = [
     "CQLTrainer",
     "DDPGTrainer",
     "DQNTrainer",
-    "GRPOOptimizationStepper",
     "GRPOTrainer",
     "IQLTrainer",
     "OfflineToOnlineTrainer",
